Swot Notes Modal is in it's own file
